Increase zone memory
Double the amount of preallocated zone memory which is needed for some
recent high resolution texture packs (e.g. "Neural Upscale" by
MamiyaOtaru).

Increasing this makes sense now that we have
https://github.com/ioquake/ioq3/commit/20a77764dd7f62101f8a3750471dab300a840858#diff-12a54363af4f3f3aaa86590d1d6b3e38
and it doesn't hurt either in times where 8GB+ of ram is _very_ common.

Additionally, as stated in
https://github.com/ioquake/ioq3/blob/master/code/qcommon/common.c#L1435
is is impossible to set com_zonemegs via config file and thus making
it hard to ship zonemegs increasement with any mods or packs.